Added TorchVis
TorchVis is a versatile tool that handles writer opening/closing for the end user, and provides one API that can accept input from both Caffe2 and PyTorch. It can write out to both Visdom and TensorBoard-compatible formats. It utilizes a pub-sub pattern in which you subscribe a desired output format and it publishes output to all currently subscribed output formats.