Added test for image/video values
We no longer assume in image() in tensorboardX/summary.py that the user will pass in an image with values in [0, 1] (i.e. np.float32 values). Instead, the code now checks the image's datatype to determine if we need to scale the image to [0, 255] or if it is already on that scale. test_summary.py has been added to test this.