Global logger (#556)
* [experimental] global multiprocess logger

This PR adds a feature that different processes can add summaries (e.g. scalar) to the same TensorBoard file. In the case of a modularized program, there is no need to pass the SummaryWriter instance around. Now we provide a logging style syntax like GlobalSummaryWriter.getSummaryWriter() to get the writer from anywhere. Both cases depend on a new GlobalSummaryWriter class, the usage is similar to the SummaryWriter except that the parameter global_step in add_x(...) shall not be passed. This is by design because the execution order of a multiprocess program is not guaranteed, the global step should be maintained by the GlobalSummaryWriter instead of determined by each child process.