Use matplotlib's agg backend in test_figure.py.
By default, matplotlib will use an interactive backend when it detects
that one is available. This is both unnecessary and, in some cases, can
break builds.

One example of a build breaking is on macOS Ventura when these tests
are run using a non-root user (such as a user used for builds) that is
not the logged-in user. The interactive macOS backend crashes with an
error in _xpc_api_misuse when creating an NSWindow, likely due to some
tighter security requirements.